requirements.txt
playwright==1.40.0
psycopg2-binary==2.9.9
PyQt5==5.15.10
Flask==2.3.3
stripe==7.8.0
python-dotenv==1.0.0
requests==2.31.0
